requirements.txt
streamlit
tiktoken
tokenizer
azure-ai-documentintelligence
azure-ai-formrecognizer
azure-identity
azure-core
azure-search-documents==11.6.0b3
python-dotenv
openai
numpy
pyodbc
num2words
matplotlib
plotly
scipy
scikit-learn
pandas
tokenizer
PrettyTable
nltk
